Split the hover module up
