+++ crates/ide/src/display.rs
@@ -2,7 +2,4 @@
 //! into types that may be used to render in a UI.
 
 pub(crate) mod navigation_target;
-
 pub(crate) use navigation_target::{ToNav, TryToNav};
-
-pub(crate) use syntax::display::macro_label;
